Fix header

+++ sort.h
@@ -24,7 +24,7 @@ void insertion_sort_list(listint_t **list);
 void selection_sort(int *array, size_t size);
 void quick_sort(int *array, size_t size);
 
-void swap(int *array, size_t i, size_t j);
+void swap(int *array, int i, int j);
 int partition(int *array, int low, int high, size_t size);
 void quickSort(int *array, int low, int high, size_t size);
